Move SampleType constants to SampleTypeAPIHelper per CR
